+++ .env
@@ -4,7 +4,9 @@ DEBUG=
 POSTGRES_DB=
 POSTGRES_USER=
 POSTGRES_PASSWORD=
+
 POSTGRES_HOST=database
+BROKER_HOST=rabbitmq
 
 ADMIN_USERNAME=
 ADMIN_EMAIL=
