+++ .github/workflows/ci-cd-master.yml
@@ -32,6 +32,7 @@ jobs:
           envkey_POSTGRES_USER: ${{ secrets.POSTGRES_USER }}
           envkey_POSTGRES_PASSWORD: ${{ secrets.POSTGRES_PASSWORD }}
           envkey_POSTGRES_HOST: database
+          envkey_BROKER_HOST: rabbitmq
           envkey_ADMIN_USERNAME: ${{ secrets.ADMIN_USERNAME }}
           envkey_ADMIN_EMAIL: ${{ secrets.ADMIN_EMAIL }}
           envkey_ADMIN_PASSWORD: ${{ secrets.ADMIN_PASSWORD }}
