+++ dockerhub.yml
@@ -13,6 +13,19 @@ services:
       - "5432:5432"
     volumes:
       - database-risks:/var/lib/postgresql/data
+    networks:
+      - database_network
+
+  rabitmq:
+    container_name: "doctorinna-risks-broker"
+    image: rabbitmq:latest
+    ports:
+      - "5672:5672"
+    depends_on:
+      - database
+    networks:
+      - database_network
+      - nginx_network
 
   django_gunicorn:
     container_name: "doctorinna-gunicorn"
@@ -23,8 +36,13 @@ services:
       - .env
     ports:
       - "8000:8000"
+    links:
+      - rabbitmq
     depends_on:
-      - database
+      - rabbitmq
+    networks:
+      - database_network
+      - nginx_network
 
   nginx:
     container_name: "doctorinna-nginx"
@@ -35,6 +53,16 @@ services:
       - "80:80"
     depends_on:
       - django_gunicorn
+    networks:
+      - nginx_network
+    links:
+      - rabbitmq
+
+networks:
+  nginx_network:
+    driver: bridge
+  database_network:
+    driver: bridge
 
 volumes:
   database-risks:
