+++ requirements.txt
@@ -1,6 +1,13 @@
+amqp==5.0.6
 asgiref==3.4.1
+billiard==3.6.4.0
+celery==5.1.2
 certifi==2021.5.30
 charset-normalizer==2.0.6
+click==7.1.2
+click-didyoumean==0.3.0
+click-plugins==1.1.1
+click-repl==0.2.0
 coreapi==2.3.3
 coreschema==0.0.4
 Django==3.2.7
@@ -14,10 +21,12 @@ idna==3.2
 itypes==1.2.0
 Jinja2==3.0.2
 joblib==1.1.0
+kombu==5.1.0
 MarkupSafe==2.0.1
 numpy==1.21.2
 openapi-codec==1.3.2
 pandas==1.3.3
+prompt-toolkit==3.0.20
 psycopg2-binary==2.9.1
 python-dateutil==2.8.2
 pytz==2021.3
@@ -32,3 +41,5 @@ text-unidecode==1.3
 threadpoolctl==3.0.0
 uritemplate==3.0.1
 urllib3==1.26.7
+vine==5.0.0
+wcwidth==0.2.5
